feat(monitoring): allow single-line instrumentation
We extend the monitoring API to allow passing in an optional
argument. This is interpreted as necessary by each event. In
particular, for the LINE event, we expect the argument to be a
dictionary, when one is given. The dictionary should contain a
"offsets" key, whose value is a set of offsets where the local
LINE event should be instrumented on.